requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.1.0
Django==5.0.1
django-crontab==0.7.1
djangorestframework==3.14.0
dnspython==2.5.0
joblib==1.3.2
kombu==5.3.5
numpy==1.26.3
pandas==2.2.0
prompt-toolkit==3.0.43
pyarrow==15.0.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
redis==5.0.1
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
xgboost==2.0.3
